Finished moving onboarding repository to it's own module
